+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Added info text for SkipDisk scenarios
+
 ### Fixed
 - parse_time_str/since argument: better wrong input handling